refactor(cli): deduplicate _find_project_dir + _resolve_local_pack_dir into _shared.py
Step 2 of Work Item #10 (audit Check 33 — duplicate logic detection).

The audit reported 2 duplicates; investigation found more:
- _find_project_dir was copy-pasted in 6 CLI modules with 3 slightly
  different error messages
- _resolve_local_pack_dir was a 35-line identical copy in 2 modules
  (ds_cmd.py, kit_cmd.py)

Consolidated both into a new private module streamtex/cli/_shared.py
(matches the existing _stx_toml.py, _toml_helpers.py, _install_local_packs.py
underscore-prefixed pattern).

Function names kept their leading underscore so all 6 call sites can
import without renaming:
    from ._shared import _find_project_dir

Canonical error message standardized to the longer/more helpful version:
"No pyproject.toml in current directory. Run from a StreamTeX project root."
(previously 3 variants existed across the 6 files).

Files updated:
- streamtex/cli/_shared.py        — NEW, 65 lines (both helpers + docstring)
- streamtex/cli/pack_cmd.py       — import; removed local copy (-9 lines)
- streamtex/cli/component_cmd.py  — import; removed local copy (-9 lines)
- streamtex/cli/ds_cmd.py         — import; removed local copies (-44 lines)
- streamtex/cli/kit_cmd.py        — import; removed local copies (-44 lines)
- streamtex/cli/dev_cmd.py        — import; removed local copy (-10 lines)
- streamtex/cli/validate_cmd.py   — import; removed local copy (-6 lines)

Net: -57 lines duplicated code removed, +65 lines centralized.

Validation:
- ruff check: clean across all 7 files
- 528 CLI tests pass, 0 failures
- Smoke import of all 6 CLI modules: clean
